New checking of generated source files

+++ src/main/java/org/gtk/gdk/DeleteEvent.java
@@ -14,20 +14,33 @@ public class DeleteEvent extends org.gtk.gdk.Event {
         Gdk.javagi$ensureInitialized();
     }
     
+    private static final java.lang.String C_TYPE_NAME = "GdkDeleteEvent";
+    
     /**
-     * Memory layout of the native struct is unknown (no fields in the GIR file).
-     * @return always {code Interop.valueLayout.ADDRESS}
+     * Memory layout of the native struct is unknown.
+     * @return always {@code Interop.valueLayout.ADDRESS}
      */
     public static MemoryLayout getMemoryLayout() {
         return Interop.valueLayout.ADDRESS;
     }
     
+    @ApiStatus.Internal
     public DeleteEvent(io.github.jwharm.javagi.Refcounted ref) {
         super(ref);
     }
     
-    /** Cast object to DeleteEvent */
+    /**
+     * Cast object to DeleteEvent if its GType is a (or inherits from) "GdkDeleteEvent".
+     * @param  gobject            An object that inherits from GObject
+     * @return                    An instance of "DeleteEvent" that points to the memory address of the provided GObject.
+     *                            The type of the object is checked with {@code g_type_check_instance_is_a}.
+     * @throws ClassCastException If the GType is not derived from "GdkDeleteEvent", a ClassCastException will be thrown.
+     */
     public static DeleteEvent castFrom(org.gtk.gobject.Object gobject) {
-        return new DeleteEvent(gobject.refcounted());
+        if (org.gtk.gobject.GObject.typeCheckInstanceIsA(gobject.g_type_instance$get(), org.gtk.gobject.GObject.typeFromName("GdkDeleteEvent"))) {
+            return new DeleteEvent(gobject.refcounted());
+        } else {
+            throw new ClassCastException("Object type is not an instance of GdkDeleteEvent");
+        }
     }
 }
